Move PGO support to Common
This part is shareable with other CorInfo implementers.
